+++ src/crm_css_maintenance.c
@@ -21,10 +21,10 @@
 //  and include the routine declarations file
 #include "crm114.h"
 
-static long crm_zapcss(FEATUREBUCKET_TYPE *h,
-        unsigned long                      hs,
-        unsigned long                      start,
-        unsigned long                      end);
+static int crm_zapcss(FEATUREBUCKET_TYPE *h,
+        unsigned int                      hs,
+        unsigned int                      start,
+        unsigned int                      end);
 
 //     How to microgroom a .css file that's getting full
 //
@@ -42,20 +42,20 @@ static long crm_zapcss(FEATUREBUCKET_TYPE *h,
 //     how the file looks, and if necessary, we get rid of some data.
 //     R is the "MICROGROOM_RESCALE_FACTOR"
 //
-long crm_microgroom(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
-        long hs, unsigned long hindex)
+int crm_microgroom(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
+        int hs, unsigned int hindex)
 {
-    long i, j, k;
-    static long microgroom_count = 0;
-    long steps;
-    long packstart;   // first used bucket in the chain
-    long packlen;     // # of used buckets in the chain
-    long packend;     // last used bucket in the chain
+    int i, j, k;
+    static int microgroom_count = 0;
+    int steps;
+    int packstart;   // first used bucket in the chain
+    int packlen;     // # of used buckets in the chain
+    int packend;     // last used bucket in the chain
     //  for stochastic grooming we need a place for the random...
-    unsigned long randy;
-    long zeroed_countdown;
-    long actually_zeroed;
-    long force_rescale;
+    unsigned int randy;
+    int zeroed_countdown;
+    int actually_zeroed;
+    int force_rescale;
 
     j = 0;
     k = 0;
@@ -70,7 +70,7 @@ long crm_microgroom(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
     {
         if (microgroom_count == 1)
             fprintf(stderr, "CSS file too full: microgrooming this css chain: ");
-        fprintf(stderr, " %ld ",
+        fprintf(stderr, " %d ",
                 microgroom_count);
     }
 
@@ -249,10 +249,10 @@ long crm_microgroom(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
 //
 //      crm_zapcss - the distance-heuristic microgroomer core.
 
-static long crm_zapcss(FEATUREBUCKET_TYPE *h,
-        unsigned long                      hs,
-        unsigned long                      start,
-        unsigned long                      end)
+static int crm_zapcss(FEATUREBUCKET_TYPE *h,
+        unsigned int                      hs,
+        unsigned int                      start,
+        unsigned int                      end)
 {
     //     A question- what's the ratio deprecation ratio between
     //     "distance from original" vs. low point value?  The original
@@ -277,20 +277,20 @@ static long crm_zapcss(FEATUREBUCKET_TYPE *h,
 #define DWEIGHT 1.0
 #define DWEIGHT2 0.0
 
-    long vcut;
-    long zcountdown;
-    unsigned long packlen;
-    unsigned long k;
-    long actually_zeroed;
+    int vcut;
+    int zcountdown;
+    unsigned int packlen;
+    unsigned int k;
+    int actually_zeroed;
 
     vcut = 1;
     packlen = end - start;
-    //  fprintf(stderr, " S: %ld, E: %ld, L: %ld ", start, end, packlen );
+    //  fprintf(stderr, " S: %d, E: %d, L: %d ", start, end, packlen );
     zcountdown = packlen / 32; //   get rid of about 3% of the data
     actually_zeroed = 0;
     while (zcountdown > 0)
     {
-        //  fprintf(stderr, " %ld ", vcut);
+        //  fprintf(stderr, " %d ", vcut);
         for (k = start; k <= end;  k++)
         {
             if (h[k].key != 0)      // key == 0 means "special- don't zero!"
@@ -322,7 +322,7 @@ static long crm_zapcss(FEATUREBUCKET_TYPE *h,
 }
 
 void crm_packcss(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
-        long hs, long packstart, long packlen)
+        int hs, int packstart, int packlen)
 {
     //    How we pack...
     //
@@ -333,7 +333,7 @@ void crm_packcss(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
     //    back up past us (since the file must contain at least one empty)
     //    and so it's still below us in the file.
 
-    //fprintf(stderr, "Packing %ld len %ld total %ld",
+    //fprintf(stderr, "Packing %d len %d total %d",
     //       packstart, packlen, packstart+packlen);
     //  if (packstart+packlen >= hs)
     //  fprintf(stderr, " BLORTTTTTT ");
@@ -349,11 +349,11 @@ void crm_packcss(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
 }
 
 void crm_packseg(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
-        long hs, long packstart, long packlen)
+        int hs, int packstart, int packlen)
 {
-    unsigned long ifrom, ito;
+    unsigned int ifrom, ito;
     crmhash_t thash, tkey;
-    unsigned long tvalue;
+    unsigned int tvalue;
     unsigned char tseen;
 
     //  keep the compiler quiet - tseen is used only if seen_features
@@ -361,7 +361,7 @@ void crm_packseg(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
     tseen = 0;
 
     if (internal_trace)
-        fprintf(stderr, " < %ld %ld >", packstart, packlen);
+        fprintf(stderr, " < %d %d >", packstart, packlen);
 
     for (ifrom = packstart; ifrom < packstart + packlen; ifrom++)
     {
@@ -411,7 +411,7 @@ void crm_packseg(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
             ito = thash % hs;
             if (ito == 0)
                 ito = 1;
-            // fprintf(stderr, "a %ld", ito);
+            // fprintf(stderr, "a %d", ito);
 
             while (!((h[ito].value == 0)
                      || (h[ito].hash == thash
@@ -420,7 +420,7 @@ void crm_packseg(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
                 ito++;
                 if (ito >= hs)
                     ito = 1;
-                // fprintf(stderr, "a %ld", ito);
+                // fprintf(stderr, "a %d", ito);
             }
 
             //
@@ -452,11 +452,11 @@ void crm_packseg(FEATUREBUCKET_TYPE *h, unsigned char *seen_features,
     }
 }
 
-int crm_create_cssfile(char *cssfile, long buckets,
-        long major, long minor, long spectrum_start)
+int crm_create_cssfile(char *cssfile, int buckets,
+        int major, int minor, int spectrum_start)
 {
     FILE *f;
-    long i;
+    int i;
     FEATUREBUCKET_STRUCT feature = { 0 };
 
     if (user_trace)
